Add non-writable error handler
